feature(Costi Notifica Fase 5)[QA-13814] Creazione di una classe AwsServiceSteps estrapolata da SharedSteps, dove verranno messi tutti i metodi che riguardano interrogazioni a DynamoDB o Cloudwatch
