+++ pom.xml
@@ -101,7 +101,6 @@
         <dependency>
             <groupId>software.amazon.awssdk</groupId>
             <artifactId>cognitoidentityprovider</artifactId>
-            <version>2.25.23</version>
         </dependency>
 
         <!-- AWS SDK v2 -->
