+++ src/test/java/it/pagopa/pn/cucumber/steps/SharedSteps.java
@@ -33,7 +33,6 @@ import it.pagopa.pn.client.b2b.pa.wrapper.RecipientWrapper;
 import it.pagopa.pn.client.web.generated.openapi.clients.externalUserAttributes.addressBook.model.CourtesyDigitalAddress;
 import it.pagopa.pn.cucumber.steps.pa.notificationVersions.NotificationStepsInterface;
 import it.pagopa.pn.cucumber.steps.pa.notificationVersions.NotificationVersion;
-import it.pagopa.pn.cucumber.steps.pa.utilityVersions.AwsUtils;
 import it.pagopa.pn.cucumber.steps.pa.utilityVersions.B2bUtils;
 import it.pagopa.pn.cucumber.steps.utilitySteps.Costanti;
 import it.pagopa.pn.cucumber.steps.utilitySteps.Destinatario;
@@ -57,11 +56,6 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.util.Base64Utils;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.HttpStatusCodeException;
-import software.amazon.awssdk.services.cloudwatchlogs.CloudWatchLogsClient;
-import software.amazon.awssdk.services.cloudwatchlogs.model.FilterLogEventsRequest;
-import software.amazon.awssdk.services.cloudwatchlogs.model.FilterLogEventsResponse;
-import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
-import software.amazon.awssdk.services.dynamodb.model.*;
 
 import java.io.IOException;
 import java.security.SecureRandom;
@@ -76,7 +70,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assumptions.assumeThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
-import static org.awaitility.Awaitility.await;
 
 
 @Scope(value = ConfigurableBeanFactory.SCOPE_PROTOTYPE)
@@ -158,9 +151,6 @@ public class SharedSteps {
     @Getter
     private final ObjectMapper objMapper;
 
-    @Getter
-    private final AwsUtils awsUtils;
-
     private boolean checkAuditLogDisabled;
 
     /**
@@ -216,8 +206,7 @@ public class SharedSteps {
                        PnGPDClientImpl pnGPDClientImpl,
                        PnPaymentInfoClientImpl pnPaymentInfoClientImpl,
                        IPnTosPrivacyClientImpl iPnTosPrivacyClientImpl,
-                       PnB2bClientTimingConfigs timingConfigs,
-                       AwsUtils awsUtils) {
+                       PnB2bClientTimingConfigs timingConfigs) {
         this.context = context;
         this.b2bClient = b2bClient;
         this.pollingFactory = pollingFactory;
@@ -230,7 +219,6 @@ public class SharedSteps {
         this.pnPaymentInfoClientImpl = pnPaymentInfoClientImpl;
         this.iPnTosPrivacyClientImpl = iPnTosPrivacyClientImpl;
         this.timingConfigs = timingConfigs;
-        this.awsUtils = awsUtils;
         this.iuvGPD = new ArrayList<>();
         this.objMapper = JsonMapper.builder().addModule(new JavaTimeModule()).build();
         this.secureRandom = new SecureRandom();
@@ -1388,77 +1376,5 @@ public class SharedSteps {
         log.info("IUN OLDER {} GG: {}", lowerLimit, oldNotification.getIun());
     }
 
-    @Then("verifico che su DynamoDB {is} presente in timeline l'elemento {string}")
-    public void checkTimelineFromDynamoDB(boolean isPresent, String timelineElement) {
 
-        Map<String, AttributeValue> expressionAttributeValues = new HashMap<>();
-        expressionAttributeValues.put(":v_iun", AttributeValue.builder().s(notificationIun).build());
-        expressionAttributeValues.put(":v_category", AttributeValue.builder().s(timelineElement).build());
-
-        DynamoDbClient dbClient = awsUtils.getDynamoDbClient();
-        QueryRequest queryRequest = AwsUtils.buildPnTimelinesCategoryRequest(expressionAttributeValues);
-        QueryResponse queryResponse = dbClient.query(queryRequest);
-
-        log.info("Elementi trovati con categoria {}: {}", timelineElement, queryResponse.count());
-        if (isPresent) {
-            assertThat(queryResponse.items().size()).as("La response non contiene nessun elemento con category " + timelineElement).isGreaterThan(0);
-
-            for (int i = 0; i < queryResponse.items().size(); i++) {
-                Map<String, AttributeValue> item = queryResponse.items().get(i);
-                log.info("--- ELEMENTO TIMELINE {} ---", i + 1);
-                item.forEach((key, value) -> {
-                    Object val = (value.s() != null) ? value.s() :
-                            (value.n() != null) ? value.n() :
-                                    (value.bool() != null) ? value.bool() : value.toString();
-                    log.info("{}: {}", key, val);
-                });
-            }
-        } else {
-            assertThat(queryResponse.items().size()).as("La response non deve contenere nessun elemento con category " + timelineElement).isEqualTo(0);
-        }
-    }
-
-    /**
-     * Dalla mappa di parametri passata in input si va a comporre la stringa con i filtri di ricerca per i log.
-     * Il nome delle chiavi impostato nel file feature è del tutto irrilevante, serve unicamente a migliorare la leggibilità.
-     * La sola eccezione è il campo iun (ignoreCase): se è impostato su auto, viene recuperato dal valore di notificationIun;
-     * in tutti gli altri casi, viene impostato col valore passato.
-     * La scrittura dei log su CloudWatch è repentina, ma possono tuttavia capitare casi in cui il log viene scritto pochi secondi dopo
-     * rispetto a quando viene eseguita la ricerca. Per evitare fail dei test dovuti a questa motivazione, il metodo riesegue la ricerca
-     * tramite polling prima di dichiarare fallito il test.
-     */
-    @And("verifico la presenza di un audit log su {string} negli ultimi {int} minuti riportante i seguenti dati nel messaggio")
-    public void checkAuditLogFromAws(String microservice, int minutes, Map<String, String> queryFiltersMap) throws InterruptedException {
-        if (!checkAuditLogDisabled) {
-            try {
-                StringBuilder sb = new StringBuilder();
-                queryFiltersMap.forEach((key, value) -> {
-                    if (key.equalsIgnoreCase("IUN")) {
-                        sb.append("\"").append(value.equals("auto") ? notificationIun : value).append("\" ");
-                    } else {
-                        sb.append("\"").append(value).append("\" ");
-                    }
-                });
-                String search = sb.toString().trim();
-                CloudWatchLogsClient cloudWatchLogsClient = awsUtils.getCloudWatchLogsClient();
-
-                await().atMost(2, TimeUnit.MINUTES).pollInterval(10, TimeUnit.SECONDS).ignoreExceptions().untilAsserted(() -> {
-                    FilterLogEventsRequest logRequest = AwsUtils.buildCloudWatchLogRequest(microservice, search, minutes);
-                    FilterLogEventsResponse logResponse = cloudWatchLogsClient.filterLogEvents(logRequest);
-                    assertThat(logResponse.events().size()).as("Non è stato trovato nessun log che soddisfi la search %s", search).isGreaterThan(0);
-                });
-            } catch (AssertionError assertionError) {
-                throwAssertionErrorWithIUN(assertionError);
-            }
-        }
-    }
-
-    /**
-     * Metodo a soli scopi di debugging per evitare che vengano eseguiti i controlli sugli audit con tempistiche obsolete
-     * (da usare solitamente in combo con il metodo che imposta lo iun e la PA di SharedSteps)
-     */
-    @Given("vengono disabilitati i check sugli audit log")
-    public void disableAuditLogCheck() {
-        checkAuditLogDisabled = true;
-    }
 }
